Add mode display control
