+++ src/main/java/com/Nxer/TwistSpaceTechnology/common/machine/multiMachineClasses/GTCM_MultiMachineBase.java
@@ -1,5 +1,7 @@
 package com.Nxer.TwistSpaceTechnology.common.machine.multiMachineClasses;
 
+import static gregtech.api.metatileentity.BaseTileEntity.TOOLTIP_DELAY;
+
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Iterator;
@@ -29,6 +31,12 @@ import com.Nxer.TwistSpaceTechnology.config.Config;
 import com.Nxer.TwistSpaceTechnology.util.TextEnums;
 import com.gtnewhorizon.structurelib.alignment.constructable.IConstructable;
 import com.gtnewhorizon.structurelib.alignment.constructable.ISurvivalConstructable;
+import com.gtnewhorizons.modularui.api.drawable.IDrawable;
+import com.gtnewhorizons.modularui.api.drawable.UITexture;
+import com.gtnewhorizons.modularui.api.widget.IWidgetBuilder;
+import com.gtnewhorizons.modularui.api.widget.Widget;
+import com.gtnewhorizons.modularui.common.widget.ButtonWidget;
+import com.gtnewhorizons.modularui.common.widget.FakeSyncWidget;
 
 import gregtech.api.gui.modularui.GTUITextures;
 import gregtech.api.interfaces.metatileentity.IMetaTileEntity;
@@ -612,6 +620,7 @@ public abstract class GTCM_MultiMachineBase<T extends GTCM_MultiMachineBase<T>>
      * Use {@link #machineMode} to get current machine mode index.
      * Override {@link #getMachineModeName(int)} to set name for each mode.
      * Override {@link #setMachineModeIcons()} to set button icon.
+     * Override {@link #setMachineMode(int)} or {@link #nextMachineMode()} to restrict mode change.
      */
     public int totalMachineMode() {
         return 1;
@@ -646,6 +655,36 @@ public abstract class GTCM_MultiMachineBase<T extends GTCM_MultiMachineBase<T>>
         return machineMode + 1;
     }
 
+    public boolean canButtonSwitchMode() {
+        return supportsMachineModeSwitch();
+    }
+
+    @Override
+    public ButtonWidget createModeSwitchButton(IWidgetBuilder<?> builder) {
+        if (!supportsMachineModeSwitch()) return null;
+        Widget button = new ButtonWidget().setOnClick((clickData, widget) -> {
+            if (canButtonSwitchMode()) {
+                onMachineModeSwitchClick();
+                setMachineMode(nextMachineMode());
+            }
+        })
+            .setPlayClickSound(supportsMachineModeSwitch())
+            .setBackground(() -> {
+                List<UITexture> ret = new ArrayList<>();
+                if (supportsMachineModeSwitch()) {
+                    ret.add(GTUITextures.BUTTON_STANDARD);
+                    ret.add(getMachineModeIcon(getMachineMode()));
+                } else return null;
+                return ret.toArray(new IDrawable[0]);
+            })
+            .attachSyncer(new FakeSyncWidget.IntegerSyncer(this::getMachineMode, this::setMachineMode), builder)
+            .addTooltip(StatCollector.translateToLocal("GT5U.gui.button.mode_switch"))
+            .setTooltipShowUpDelay(TOOLTIP_DELAY)
+            .setPos(getMachineModeSwitchButtonPos())
+            .setSize(16, 16);
+        return (ButtonWidget) button;
+    }
+
     @Override
     public void onScrewdriverRightClick(ForgeDirection side, EntityPlayer aPlayer, float aX, float aY, float aZ) {
         if (getBaseMetaTileEntity().isServerSide()) {
@@ -658,11 +697,15 @@ public abstract class GTCM_MultiMachineBase<T extends GTCM_MultiMachineBase<T>>
         }
     }
 
+    public boolean showModeInWaila() {
+        return supportsMachineModeSwitch();
+    }
+
     @Override
     public void getWailaNBTData(EntityPlayerMP player, TileEntity tile, NBTTagCompound tag, World world, int x, int y,
         int z) {
         super.getWailaNBTData(player, tile, tag, world, x, y, z);
-        if (supportsMachineModeSwitch()) {
+        if (showModeInWaila()) {
             tag.setInteger("mode", machineMode);
         }
     }
@@ -671,8 +714,8 @@ public abstract class GTCM_MultiMachineBase<T extends GTCM_MultiMachineBase<T>>
     public void getWailaBody(ItemStack itemStack, List<String> currentTip, IWailaDataAccessor accessor,
         IWailaConfigHandler config) {
         super.getWailaBody(itemStack, currentTip, accessor, config);
-        if (supportsMachineModeSwitch()) {
-            final NBTTagCompound tag = accessor.getNBTData();
+        final NBTTagCompound tag = accessor.getNBTData();
+        if (tag.hasKey("mode")) {
             currentTip.add(
                 StatCollector.translateToLocal("GT5U.machines.oreprocessor1") + " "
                     + EnumChatFormatting.WHITE
